simplify qr2 driver
